Split GetMediaChannel into two functions.
It splits GetMediaChannel into EstablishMediaChannelSync and GetMediaChannel
how it was supposed to be originally. Otherwise sometime it tries
to recreate the channel after a media player release
which causes weird bugs.

The fix seems to help enormously with overall stability and removes
GL context related crashes when the page is getting closed.

Some more testing will be required to confirm though.